### 2020-07-18 #87 @whedon
[PRE REVIEW]: Interactive Bin Packing: A Java Application for Learning Constructive Heuristics for Combinatorial Optimization
**Submitting author:** @cicirello (<a href="http://orcid.org/0000-0003-1072-8559">Vincent A Cicirello</a>)
**Repository:** <a href="https://github.com/cicirello/InteractiveBinPacking" target ="_blank">https://github.com/cicirello/InteractiveBinPacking</a>
**Version:** v3.0

@cicirello if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/jose-reviewers) have already agreed to review for JOSE and may be suitable for this submission.
