### 2021-02-20 #87 @labarba
Hi @cicirello 👋  — I apologize for not delivering on expectations here, but JoSE became abandoned when I got crushed with other commitments, plus online teaching, and many changes and stresses due to the world chaos of the past year. I'm just getting back to audit all the dropped balls and try to pick them up one by one.

I have read your paper, and understand that your submission is about a tutorial on the bin-packing optimization problem. However, I am not able to figure out how your tutorial works, without presumably installing the Java application? Can you help me understand how it works?
