### 2021-02-20 #87 @cicirello
@labarba I totally understand about the delay. The application itself explains how it functions, etc. There is a prebuilt executable jar file in the dist directory. Assuming you have Java 8 or higher on your system, you can run it by either double clicking the jar file or executing the following on the command line: `java -jar interactive-bin-packing-3.0.jar`

Once the application is open, there is a Tutorial option in the Info menu. That will walk you through usage of all application functions as well as the tutorial itself.
