### 2021-02-20 #87 @labarba
Thanks!

I, personally, am not a computer scientist, have never touched Java, and cannot install the application. Before I can determine who might be an editor to handle this submission, I need to understand what it does. Could you briefly explain it to me, understanding that I cannot install it?
