### 2021-02-22 #87 @csev
At the request of the EiC, I have downloaded and installed this software.   Following the documentation it starts and runs and makes sense given the problem that it solves.   It feels somewhat narrow in its application - it would be a great interactive part of a week of an advanced algorithms course towards the end of undergraduate program or at the beginning of a masters program in Computer Science.    The source code is readable - I assume students looking at the source code would be part of any exercise.  It is pretty clear what it is and how it works - I am curious as to how this is used in a course.
