### 2021-02-23 #87 @cicirello
@csev I've used it in a couple different courses in a couple different ways. Here are a couple examples:

1) In an upper level undergrad AI course, just before covering randomized heuristics, I have the students work through the tutorial as part of a homework assignment (I usually give them a specific instance number so they can submit the various heuristic solutions as proof they worked through the tutorial). They usually then must work through the heuristic solutions to a problem instance not in the app.

A followup assignment involves implementing the heuristics.

I then cover in class some algorithms that rely on randomizing constructive heuristics such as ant colony optimization.

2) In that same course, I have also had students design and implement their own heuristic for the problem, and they seem to find it useful using the application to try out ideas by hand while they are designing their heuristic but befire implementing it.

3) In a non-majors intro course where students have very little, if any, prior cs background (a course with no prereqs), I have also used it in a similar way to #1 above, but without the followup implementation assignment. But for non-majors I usually provide an in class overview, whereas in #1 above the assignment to use the tutorial serves as the intro where I then cover the more advanced content in class.

4) A much earlier version lacked the functionality to verify the users' understanding of heuristic. I used that prior version in a classroom as a visualization tool while explaining the concepts. It can still be used in that way.
