+++ .github/workflows/ant.yml
@@ -16,9 +16,12 @@ jobs:
 
     steps:
     - uses: actions/checkout@v2
+
     - name: Set up JDK 1.8
       uses: actions/setup-java@v2
       with:
-        java-version: 1.8
+        distribution: 'adopt'
+        java-version: '8'
+
     - name: Build with Ant
       run: ant -noinput -buildfile build/build.xml
