+++ tests/README.md
@@ -1,29 +1,0 @@
-# Testing the Interactive Bin Packing Application
-
-## Automated Tests via JUnit
-
-This directory contains JUnit tests that cover nearly all of the Interactive Bin Packing Application, including automating the testing of nearly all of the GUI.  All non-GUI functionality is tested via JUnit tests.  The JUnit tests for the GUI simulates clicking on buttons, menu items, etc via the doClick() method, supported by most Swing elements.  In this way GUI elements are unit tested individually, as well as in combination, programmatically without the need for human tester interaction.
-
-You can execute the tests during the build process.  To include execution of the JUnit tests during the build process, from the command line, execute `ant test` (if your working directory is the build directory) or `ant -f build/build.xml test` (if your working directory is the parent of build).
-
-## Limited Manual Testing
-
-Although nearly all of the GUI is tested via JUnit, there are a few GUI elements for which automated testing is not particularly feasible.  Here is a list of these things, and what to look for.
-* _Info Menu_
-    * Click the About menu item. This should open the dialog with the copyright and license notices. It should be modal, so confirm that you cannot switch back to the app. Verify that the hyperlinks either open a web browser, or gives an appropriate error in the event your system doesn't allow this behavior. Close the About dialog.
-	* Click the Tutorial menu item. This should open a modeless dialog, with a self-guided tutorial that walks the user through explanations of combinatorial optimization, the bin packing problem, constructive heuristics, and heuristics for bin packing. Confirm that you can switch back to the application without closing the Tutorial window. Confirm that scrolling works correctly. There are hyperlinks within the Tutorial window that should scroll the window to the relevant content. Confirm that these enable navigation within the window. There are a couple of external hyperlinks (should be obvious from their text) that should either open your web browser or display an appropriate error message if your system doesn't allow the application to open a browser.  Close the Tutorial window.
-	* Click the Help menu item. This should open a modeless dialog, containing user documentation of all application functionality.  Confirm that you can switch back to the application without closing the Help window. Confirm that scrolling works correctly. There are hyperlinks within the Help window that should scroll the window to the relevant content. Confirm that these enable navigation within the window. Close the Help window.
-* _Problem Menu_
-    * Click "Select Instance Number" (the last menu item in this menu). Attempt to put something other than an integer, and click OK. It should prompt you again. Enter any integer, and click "cancel." The dialog should close and nothing should have changed. Click "Select Instance Number" again, but this time enter any integer and click "OK". The item sizes in the list of items in the main window should have changed. If they did not, then try again with a different integer (it is unlikely, but it is possible that you may have coincidentally entered the very seed used for the default instance).
-	* Move at least 1 item into a bin. Click Click "Select Instance Number" in the Problem menu, enter an integer, and click "OK." All bins should now be empty, including the one where you placed an item.  And the list of items should have new random sizes (unless you entered the same integer as previously).
-	* None of the other menu items in the Problem Menu need to be tested manually, as their tests are automated via JUnit.
-* _Mode Menu_
-    * Note: System behavior for correct decisions for each mode is already tested via JUnit, as is the code that checks if an item/bin combination is consistent with the chosen mode. What needs to be tested manually is that dialog messages appear with explanations of why user actions are not correct for a chosen mode. So the steps below are focused on deliberately making mistakes to cause the dialog windows to be opened. 
[…]
