+++ .github/dependabot.yml
@@ -1,12 +1,12 @@
-# To get started with Dependabot version updates, you'll need to specify which
-# package ecosystems to update and where the package manifests are located.
-# Please see the documentation for all configuration options:
-# https://help.github.com/github/administering-a-repository/configuration-options-for-dependency-updates
-
 version: 2
 updates:
-  - package-ecosystem: "github-actions" # See documentation for possible values
+  - package-ecosystem: "github-actions" 
     directory: "/" # Location of package manifests
-    target-branch: "development"
+    target-branch: "master"
+    schedule:
+      interval: "daily"
+  - package-ecosystem: "maven" 
+    directory: "/" # Location of package manifests
+    target-branch: "master"
     schedule:
       interval: "daily"
