+++ .github/workflows/build.yml
@@ -26,8 +26,6 @@ jobs:
     - name: Generate Jacoco Badge
       id: jacoco
       uses: cicirello/jacoco-badge-generator@v2
-      with:
-        generate-branches-badge: true
 
     - name: Log coverage percentage
       run: |
@@ -39,3 +37,20 @@ jobs:
       with:
         name: jacoco-report
         path: target/site/jacoco/
+
+    - name: Create pull request
+      if: ${{ github.event_name == 'push' }}
+      uses: peter-evans/create-pull-request@v3.9.2
+      with:
+        title: "Autogenerated Jacoco Coverage Badge"
+        body: > 
+          Autogenerated Jacoco coverage badge, generated by 
+          the [jacoco-badge-generator](https://github.com/cicirello/jacoco-badge-generator) 
+          GitHub action.
+        commit-message: "Autogenerated Jacoco Coverage Badge"
+        author: Vincent A Cicirello <cicirello@users.noreply.github.com>
+        committer: Vincent A Cicirello <cicirello@users.noreply.github.com>
+        branch: create-pull-request/badge
+        delete-branch: true
+        labels: |
+          automated
