### 2021-05-17 d731a14
Added missing check for invalid mode
