+++ .github/workflows/build.yml
@@ -26,6 +26,7 @@ jobs:
     - name: Generate Jacoco Badge
       id: jacoco
       uses: cicirello/jacoco-badge-generator@v2
+        generate-branches: true
 
     - name: Log coverage percentage
       run: |
