+++ .github/workflows/build.yml
@@ -27,7 +27,7 @@ jobs:
       id: jacoco
       uses: cicirello/jacoco-badge-generator@v2
       with:
-        generate-branches: true
+        generate-branches-badge: true
 
     - name: Log coverage percentage
       run: |
