+++ src/test/java/org/cicirello/ibp/GUITestCases.java
@@ -547,6 +547,194 @@ public class GUITestCases {
 		}
 	}
 	
+	@Test
+	public void testBottomPanelFF() {
+		int[] sizes = { 7, 2, 8, 5};
+		//int[] sizes2 = { 6, 3, 18, 4, 1 };
+		CallBack cb = new CallBack() {
+			@Override public void call() { }
+		};
+		int numBins = 3;
+		class CallCount {
+			int moveCount;
+		}
+		final CallCount cc = new CallCount();
+		Floor floor = new Floor(sizes);
+		ApplicationState state = new ApplicationState(numBins, floor, cb, cb, cb);
+		BottomPanelTestVersion bottom = new BottomPanelTestVersion(null, state, new CallBack() { 
+			@Override public void call() { cc.moveCount++; }
+		});
+		Component[] c = bottom.getComponents();
+		JPanel movePanel = (JPanel)c[0];
+		@SuppressWarnings("unchecked") 
+		JComboBox<Item> move = (JComboBox<Item>)movePanel.getComponent(1);
+		JPanel destPanel = (JPanel)c[1];
+		@SuppressWarnings("unchecked")
+		JComboBox<Bin> to = (JComboBox<Bin>)destPanel.getComponent(1);
+		JButton mButton = (JButton)c[2];
+		
+		state.setMode(ApplicationState.MODE_FIRST_FIT);
+		bottom.setExpectedModeString("First-Fit Mode");
+		move.setSelectedIndex(1);
+		to.setSelectedIndex(0);
+		mButton.doClick();
+		assertTrue(bottom.didErrorMessage());
+		bottom.resetError();
+
+		move.setSelectedIndex(0);
+		to.setSelectedIndex(0);
+		mButton.doClick();
+		assertFalse(bottom.didErrorMessage());
+		
+		move.setSelectedIndex(0);
+		to.setSelectedIndex(1);
+		mButton.doClick();
+		assertTrue(bottom.didErrorMessage());
+		bottom.resetError();
+	}
+	
+	@Test
+	public void testBottomPanelModesFFD() {
+		int[] sizes = { 7, 2, 8, 5};
+		int[] sizes2 = { 6, 3, 18, 4, 1 };
+		CallBack cb = new CallBack() {
+			@Override public void call() { }
+		};
+		int numBins = 3;
+		class CallCount {
+			int moveCount;
+		}
+		final CallCount cc = new CallCount();
+		Floor floor = new Floor(sizes);
+		ApplicationState state = new ApplicationState(numBins, floor, cb, cb, cb);
+		BottomPanelTestVersion bottom = new BottomPanelTestVersion(null, state, new CallBack() { 
+			@Override public void call() { cc.moveCount++; }
+		});
+		Component[] c = bottom.getComponents();
+		JPanel movePanel = (JPanel)c[0];
+		@SuppressWarnings("unchecked") 
+		JComboBox<Item> move = (JComboBox<Item>)movePanel.getComponent(1);
+		JPanel destPanel = (JPanel)c[1];
+		@SuppressWarnings("unchecked")
+		JComboBox<Bin> to = (JComboBox<Bin>)destPanel.getComponent(1);
+		JButton mButton = (JButton)c[2];
+		
+		state.setMode(ApplicationState.MODE_FIRST_FIT_DECREASING);
+		bottom.setExpectedModeString("First-Fit Decreasing Mode");
+		move.setSelectedIndex(1);
+		to.setSelectedIndex(0);
+		mButton.doClick();
+		assertTrue(bottom.didErrorMessage());
+		bottom.resetError();
+		
+		move.setSelectedIndex(0);
+		to.setSelectedIndex(0);
+		mButton.doClick();
+		assertTrue(bottom.didErrorMessage());
[…]
