### 2021-05-18 a4ce329
refactor to support testing
