+++ src/test/java/org/cicirello/ibp/GUITestCases.java
@@ -76,6 +76,64 @@ public class GUITestCases {
 	}
 	
 	@Test
+	public void testUIInit() {
+		UI ui = new UI(null);
+		ApplicationState state = ui.getApplicationState();
+		assertEquals(ApplicationState.MODE_PRACTICE, state.getMode());
+		assertEquals("Practice Mode", state.getModeString());
+		assertEquals("practice", state.getModeName());
+		Floor f = state.getFloor();
+		ArrayList<Item> initialFloorItems = f.getContents();
+		assertEquals(20, initialFloorItems.size());
+		ArrayList<Bin> bins = state.getBins();
+		assertEquals(9, bins.size());
+		for (Bin b : bins) {
+			assertEquals(100, b.capacity());
+			assertEquals(0, b.used());
+			assertTrue(b.isEmpty());
+		}
+		ArrayList<Item> items = state.getItems();
+		assertEquals(20, items.size());
+		for (Item i : items) {
+			assertTrue(f.contains(i));
+		}
+		
+		int[] sizes2 = {15, 5, 25, 10, 20};
+		int[] sorted = {5, 10, 15, 20, 25};
+		Floor f2 = new Floor(sizes2);
+		state.setNewInstance(f2);
+		f = state.getFloor();
+		ArrayList<Item> floorItems = f.getContents();
+		assertEquals(sizes2.length, floorItems.size());
+		bins = state.getBins();
+		assertEquals(9, bins.size());
+		for (Bin b : bins) {
+			assertEquals(100, b.capacity());
+			assertEquals(0, b.used());
+			assertTrue(b.isEmpty());
+		}
+		items = state.getItems();
+		assertEquals(sizes2.length, items.size());
+		int j = 0;
+		for (Item i : items) {
+			assertTrue(f.contains(i));
+			assertEquals(sizes2[j], i.size());
+			j++;
+		}
+		
+		state.sortFloor(false);
+		floorItems = f.getContents();
+		for (int i = 0; i < floorItems.size(); i++) {
+			assertEquals(sorted[i], floorItems.get(i).size());
+		}
+		state.sortFloor(true);
+		floorItems = f.getContents();
+		for (int i = 0; i < floorItems.size(); i++) {
+			assertEquals(sorted[sorted.length-i-1], floorItems.get(i).size());
+		}
+	}
+	
+	@Test
 	public void testModeMenu() {
 		int[] sizes = { 7, 2, 18, 3, 6 };
 		CallBack cb = new CallBack() {
