### 2021-05-18 ea4377a
refactor to support testing
