### 2021-05-18 95c9055
disable portion of testcase that requires a display
