### 2021-05-19 751e593
Create manual-docker.yml
