### 2021-05-19 b870f61
Update manual-docker.yml
