### 2021-05-19 d30a589
Update manual-docker.yml
