### 2021-05-19 b0f489d
Update manual-docker.yml
