+++ .github/workflows/manual-docker.yml
@@ -24,12 +24,11 @@ jobs:
           TAGS="${VERSION},${MINOR},${MAJOR}"
           echo ::set-output name=tags::${TAGS}
 
-      - name: Set up JDK 11 for ghcr.io
+      - name: Set up JDK 11
         uses: actions/setup-java@v2
         with:
           distribution: 'adopt'
           java-version: '11'
-          #server-id: ghcr.io
 
       - name: Login to DockerHub
         uses: docker/login-action@v1 
@@ -46,20 +45,6 @@ jobs:
 
       - name: Publish to GitHub Container Registry
         run: mvn compile jib:build -Djib.to.tags=${{ steps.prep.outputs.tags }} -Pghcr 
-        #env:
-          #GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }} 
-
-      #- name: Set up JDK 11 for docker.io
-        #uses: actions/setup-java@v2
-        #with:
-          #distribution: 'adopt'
-          #java-version: '11'
-          #server-id: docker.io
-          #server-username: DOCKER_USERNAME
-          #server-password: DOCKER_TOKEN
 
       - name: Publish to Docker Hub
         run: mvn compile jib:build -Djib.to.tags=${{ steps.prep.outputs.tags }} -Pdocker
-        #env:
-          #DOCKER_USERNAME: ${{ secrets.DOCKER_USERNAME }}
-          #DOCKER_TOKEN: ${{ secrets.DOCKER_PASSWORD }}
