### 2021-05-23 #87 @labarba
Hi @cicirello 👋  — we discussed your submission a bit more and find it to be too small / too narrow to merit a journal publication. Bear in mind that JOSE is evaluating the educational resources submitted with the goal of giving scholarly credit to this often unrecognized work. We want readers, authors (and their employers) to recognize JOSE papers similarly to research publications, and thus it should represent substantial scholarly effort.
