### 2021-05-23 #87 @cicirello
@labarba can you please define what is meant by a course "module" in this statement in the journal's scope: "The ideal submission size is a course module"?
