### 2021-05-23 #87 @labarba
Sure. We understand that to mean a section of a course, with its own learning objectives, content, and activities. More than a single lesson, but less than a semester course (although some submissions are a full course). The ideal module length might be around 3 to 5 weeks of a semester, or something that can be taught in an intense two-day Summer school.
