### 2021-05-23 #87 @cicirello
@labarba thanks. It might be helpful to potential authors if the 3 to 5 weeks of a course is mentioned in the scope, perhaps in parentheses after the word module. My submission otherwise has "its own learning objectives, content, and activities" but is more around 1 week to maybe 2 weeks (if students are more introductory level). It would also be helpful if determinations like this are made sooner than 10 months in "pre-review."
