### 2021-05-23 #87 @labarba
What I quote is the ideal length from my perspective, given that JOSE will give "one publication credit" for the work. We want to be mindful that JOSE papers are taken seriously by the community. One to 2 weeks of a course is 2.5 to 5 hours of instruction. That is borderline, in my assessment. We _have_ reviewed submissions that are like a tech-conference tutorial (which tend to vary between 4 and 8 hours of live instruction). It sounds like it might be possible to consider this submission, maybe if you were willing to put in a bit of work on making it more clear that this is not for a single in-class activity (that would be below the bar). Maybe combining the _application_ with a written tutorial. Did you submit to the software, or to the learning module article type?

---

Our documentation is not perfect, but like for any journal, looking at past published papers should help gage the scoping question. As for the 10 months in pre-review, JOSE is an initiative led by the volunteer editors, who have faced immeasurable challenges the past 14 months, for obvious reasons. One in five of our editors have completely stopped responding, and we don't know about their wellbeing. I personally have a work backlog measured in months not weeks, but I'm here on a Sunday to give just a bit more of me.
