+++ .github/workflows/build.yml
@@ -42,7 +42,7 @@ jobs:
 
     - name: Create pull request
       if: ${{ github.event_name == 'push' }}
-      uses: peter-evans/create-pull-request@v3.9.2
+      uses: peter-evans/create-pull-request@v3.10.0
       with:
         title: "Autogenerated Jacoco Coverage Badge"
         body: > 
