+++ .github/workflows/maven-publish.yml
@@ -55,32 +55,10 @@ jobs:
       env:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }} 
 
-    - name: Get the upload URL for a release
-      id: get
-      uses: bruceadams/get-release@v1.2.2
+    - name: Upload jar files to release as release assets
+      run: |
+        TAG=${GITHUB_REF/refs\/tags\//}
+        gh release upload ${TAG} target/${{ env.artifact_name }}-${{ steps.get_version.outputs.VERSION }}.jar
+        gh release upload ${TAG} target/${{ env.artifact_name }}-${{ steps.get_version.outputs.VERSION }}-sources.jar
       env:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-
-    - name: Upload Release Asset jar
-      id: upload-release-asset-jar 
-      uses: actions/upload-release-asset@v1.0.2
-      env:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-      with:
-        upload_url: ${{ steps.get.outputs.upload_url }} 
-        asset_path: target/${{ env.artifact_name }}-${{ steps.get_version.outputs.VERSION }}.jar
-        asset_name: ${{ env.artifact_name }}-${{ steps.get_version.outputs.VERSION }}.jar
-        asset_content_type: application/java-archive
-
-    - name: Upload Release Asset sources
-      id: upload-release-asset-src 
-      uses: actions/upload-release-asset@v1.0.2
-      env:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-      with:
-        upload_url: ${{ steps.get.outputs.upload_url }} 
-        asset_path: target/${{ env.artifact_name }}-${{ steps.get_version.outputs.VERSION }}-sources.jar
-        asset_name: ${{ env.artifact_name }}-${{ steps.get_version.outputs.VERSION }}-sources.jar
-        asset_content_type: application/java-archive
-
-
