### 2021-06-05 #87 @labarba
hi @cicirello — could you add in the paper a description of the expected instruction time with this tutorial? Include an idea of any direct instruction (lecture) needed to get the lesson started, and expected student self-directed-learning time. (I've made a note to add this to the documentation as something we want in the paper.) Also useful would be a few sentences of advice to any instructor who wished to adopt this in their own teaching.

From our [docs](https://openjournals.readthedocs.io/en/jose/submitting.html#submission-requirements):

> Computational learning modules should be complete and immediately usable for self-learning or adoption by other instructors.

We need a bit o help understanding how your submission satisfies this criterion.
