+++ assignments/README.md
@@ -1,0 +1,58 @@
+# Example Assignments
+
+This directory contains example assignments that can be
+used within the context of a course, or by individuals for
+self-learning. This directory contains two subdirectories:
+* [src](src): The source files of the assignments, which at the present
+  time are LaTeX files to make it easy for other instructors
+  to customize to their needs, and which also makes it easy to
+  provide comments intended only for instructors. So if you are a
+  course instructor, I recommend reading the comments within the
+  LaTeX files.
+* [dist](dist): Versions (as pdf files) that are ready for 
+  distribution to students if you don't have reason to modify the 
+  assignments. If you are a self-learner and not part of a class, then
+  these may be of greater interest to you than the src directory.  
+
+## License for Example Assignments
+
+As noted in the README at the top level of the repository,
+as well as in the Interactive Bin Packing application, its
+source, files, etc, the Interactive Bin Packing application
+itself is licensed under
+the [GPL-3.0](https://www.gnu.org/licenses/gpl-3.0.en.html).
+However, the GPL license is designed for application to 
+software, and as such doesn't lend itself well to licensing
+documents such as the example homework assignments.
+
+The example homework assignments contained in this directory
+are licensed to you under 
+the [CC BY-NC-SA (Creative Commons Attribution-NonCommercial-ShareAlike)](https://creativecommons.org/licenses/by-nc-sa/4.0/). Why this license?
+* It allows you to adapt the assignments to the needs of your 
+  course. The example assignments would be rather useless otherwise,
+  since what may work well for my course may work better for you if
+  you change things here or there.
+* It allows you to share. Example assignments would likewise be rather
+  useless if you couldn't share them such as with your students.
+* It requires attribution. As an academic, I feel that everything you use
+  that was created by someone else should be properly attributed.
+* NonCommercial: Well, this allows educators (or self-learning students)
+  to use for educational purposes, but prevents resale as part of a 
+  commercial product or other commercial use (e.g., prevents textbook
+  publishers from including it as part of their books, etc).
+* ShareAlike: If you change the assignments in some way and want to
+  distribute your changed version, then you must use the same license.
+  This keeps derivatives open.
+
+## Note to Instructors
+
+If you do create variations of the assignments that you think would
+be useful to others, or if you create brand new assignments that utilize
+the Interactive Bin Packing application in some way, then consider
+contributing them directly to this project, by doing the following:
+1. Fork the repository.
+2. Create a branch.
+3. Add the assignment files to the assignments directory.
+4. Submit a pull request.
+  
+
