### 2021-06-09 6641493
Added example assignment
