+++ assignments/1/assignment.tex
@@ -1,0 +1,342 @@
+% Copyright 2021 Vincent A. Cicirello
+%
+% License: CC BY-NC-SA (Creative Commons Attribution-NonCommercial-ShareAlike)
+% Full license text link: https://creativecommons.org/licenses/by-nc-sa/4.0/
+% Short summary of license: Allows remixing and adapting non-commercially, 
+% requires attribution, and adaptations required to be licensed under 
+% identical terms.
+%
+% Instructors: If you are an instructor of a course using this assignment
+% keep an eye out for comments addressed to you throughout this file.
+
+\documentclass[11pt,letterpaper]{article}
+
+\usepackage{times}
+\usepackage{url}
+\usepackage{graphicx}
+\usepackage{hyperref}
+\usepackage{enumitem}
+
+\setlength{\textwidth}{6.5in}
+\setlength{\oddsidemargin}{0in}
+\setlength{\evensidemargin}{0in}
+\setlength{\topmargin}{0in}
+\setlength{\headheight}{0in}
+\setlength{\headsep}{0in}
+\setlength{\textheight}{9in}
+
+\begin{document}
+
+\title{Combinatorial Optimization, the Bin Packing Problem, 
+and Constructive Heuristics}
+
+% Instructors: You might use the author field for the name
+% of the course or perhaps the assignment number. I've
+% left the author blank for now.
+\author{}
+
+% Instructors: You might insert the date of the assignment
+% here. I've left the date blank for now.
+\date{}
+
+\maketitle
+
+% Instructors: You might want to insert a section with your
+% own special instructions, such as how to submit the assignment,
+% deadline, etc.
+
+\section{Objectives, Prerequisites, and Time Requirements}\label{sec:objectives}
+This assignment will begin with a self-guided tutorial 
+(Section~\ref{sec:tutorial}) that will introduce you to combinatorial 
+optimization. This includes learning about what combinatorial optimization
+problems are in general, as well as learning about a specific 
+combinatorial optimization problem, the {\em Bin Packing Problem},
+used by the tutorial to illustrate concepts such as lower bounds
+as well as how to use heuristics to quickly compute satisficing
+(though not necessarily optimal) solutions. The content of the tutorial
+is entirely within a Java application (instructions for downloading
+and running are found in Section~\ref{sec:download}).
+
+After you finish working through the tutorial (Section~\ref{sec:tutorial}),
+you will work through four exercises 
+(Sections~\ref{sec:ff},~\ref{sec:ffd},~\ref{sec:bf}, and~\ref{sec:bfd}) 
+to test your knowledge of the constructive heuristics that you learned 
+about for the bin packing problem.
+
+\paragraph*{Objectives:} The objectives of this assignment include:
+\begin{itemize}[leftmargin=*, parsep=0pt, itemsep=2pt, topsep=2pt]
+\item gaining a general understanding of combinatorial optimization problems;
+\item gaining an understanding of the bin packing problem, its 
+applications, and how it is an example of a combinatorial 
+optimization problem;
+\item learning about lower bounds;
+\item learning about constructive heuristics; and
+\item learning about the most common constructive heuristics 
+for the bin packing problem, including first-fit, best-fit, 
+first-fit decreasing, and best-fit decreasing.
+\end{itemize}
+
+\paragraph*{Prerequisite Knowledge:} The tutorial does not assume
+any prior knowledge of combinatorial optimization. However, it does 
+assume that you have an understanding of introductory discrete mathematics,
+such as set theory and functions.
+
+\paragraph*{Technical Prerequisites:} This assignment does not require
+programming. However, you will need to be able to run a Java application.
+The Java application in question requires Java 11 or later. You do not need
[…]
