### 2021-06-09 cicirello/InteractiveBinPacking#26 @cicirello
Add a few example assignments utilizing the application
Add example assignments that use the Interactive Bin Packing application, and which can either be used by instructors within context of the course, or by self-learners. Include both ready-to-use pdf files, as well as LaTeX source files.  Self-learners can just go straight to the pdfs, while instructors can customize to their specific course needs using the LaTeX files.

openjournals/jose-reviews#87
