### 2021-06-09 cicirello/InteractiveBinPacking#27 @cicirello
Updates to JOSE paper based on editor feedback
Update JOSE paper based on feedback from the editor (@labarba), including the following:
1. add in the paper a description of the expected instruction time with this tutorial,
2. include an idea of any direct instruction (lecture) needed to get the lesson started,
3. expected student self-directed-learning time, and
4. also useful would be a few sentences of advice to any instructor who wished to adopt this in their own teaching.

openjournals/jose-reviews#87
