+++ paper/paper.md
@@ -20,74 +20,73 @@ bibliography: paper.bib
 
 # Summary
 
-The Interactive Bin Packing Application provides a self-guided tutorial on
-combinatorial optimization, the bin packing problem, and constructive 
-heuristics. It also enables the user to interact with bin packing instances 
-to explore their own problem solving strategies, or to test their knowledge 
-of the various constructive heuristics covered by the tutorial. The application 
-is not a solver for bin packing, but rather it is a tool for learning about 
-the bin packing problem, as well as for learning about heuristic techniques 
+Interactive Bin Packing provides a self-guided tutorial on combinatorial 
+optimization, the bin packing problem, and constructive heuristics. It also 
+enables the user to interact with bin packing instances to explore their 
+own problem solving strategies, or to test their knowledge of the various 
+constructive heuristics covered by the tutorial. The application is not a 
+solver for bin packing, but rather it is a tool for learning about the 
+bin packing problem, as well as for learning about heuristic techniques 
 for solving instances of the problem.
 
 The repository (https://github.com/cicirello/InteractiveBinPacking) contains 
-the source code, documentation, and example assignments for use by course 
-instructors or self-directed learners. An executable jar of the application can
-be downloaded from GitHub Releases as well as from Maven Central.
+the source code, documentation, and example assignments suitable for courses
+on discrete mathematics, algorithms, or artificial intelligence, or by 
+self-directed learners. An executable jar of the application can be 
+downloaded from GitHub Releases or from Maven Central.
 
-# The Bin Packing Problem
+# Bin Packing
 
-The bin packing problem is a combinatorial optimization problem in 
-which we must pack a set of items into a set of bins while minimizing 
-the number of bins used. This problem has many practical applications, 
-including obvious logistics applications, such as packing shipments 
-into trucks or airplanes, as well as many others such as creating 
-compilations of musical works, scheduling advertisements on television, 
-placing files in fixed-length memory blocks on a file system, wireless 
-network channel scheduling, job scheduling, among many 
+Bin packing is a combinatorial optimization problem in which we must 
+pack a set of items into a set of bins while minimizing the number of 
+bins used. This problem has many practical applications, including 
+within logistics, such as packing shipments into trucks or airplanes, 
+as well as many others such as creating compilations of musical works, 
+scheduling television advertisements, placing files in fixed-length 
+memory blocks on a file system, wireless network channel scheduling, 
+job scheduling, among many 
 others [@Leinberger1999; @Malkevitch2004a; @Malkevitch2004b; @Wang2006].
 
 As an NP-Hard [@Garey1979] optimization problem, it is unlikely that any 
-algorithm exists capable of optimally solving the bin packing problem in 
-polynomial time. Thus, heuristic algorithms, as well as metaheuristics, 
-are often used. Among the topics covered by this tutorial application are 
-constructive heuristics. A heuristic is a practical problem solving approach 
-that quickly provides a solution of sufficient quality, while using 
-relatively little time. A constructive heuristic is a type of heuristic 
-that iteratively applies some rule to build upon an initially empty solution 
-until it is complete. The result may or may not be optimal, and in most 
-cases you won't know whether or not it is optimal. The application guides 
-the user into learning about common constructive heuristics for bin packing.
+polynomial time algorithm exists guaranteed to optimally solve bin packing. 
+Thus, heuristic algorithms, and metaheuristics, are often used. Among the 
+topics covered by this tutorial application are constructive heuristics. A 
+heuristic is a practical problem solving approach that quickly provides a 
+solution of sufficient quality, while using relatively little time. A 
+constructive heuristic is a type of heuristic that iteratively applies some 
+rule to build upon an initially empty solution until it is complete. The 
[…]
