+++ assignments/1/assignment.tex
@@ -6,6 +6,9 @@
 % requires attribution, and adaptations required to be licensed under 
 % identical terms.
 %
+% Original version of this assignment can be found in the assignments directory
+% in this GitHub repository: https://github.com/cicirello/InteractiveBinPacking
+%
 % Instructors: If you are an instructor of a course using this assignment
 % keep an eye out for comments addressed to you throughout this file.
 
@@ -104,8 +107,11 @@ just prior to coverage of local search algorithms.
 which you will download, is open source and licensed under the GPLv3 
 (\url{https://www.gnu.org/licenses/gpl-3.0.en.html}). This assignment,
 which utilizes the application, is licensed under the CC BY-NC-SA
-(\url{https://creativecommons.org/licenses/by-nc-sa/4.0/}). 
-
+(\url{https://creativecommons.org/licenses/by-nc-sa/4.0/}). The original
+version of this assignment is available 
+at \url{https://github.com/cicirello/InteractiveBinPacking}. Students 
+should rely on the copy of the assignment provided them by their instructor
+as they may have adapted it to your course. 
 
 \section{Download Interactive Bin Packing (1--5 minutes)}\label{sec:download}
 In this assignment, we'll be using a Java application 
