### 2021-06-10 90cc223
Added SessionLog
Currently logs mode changes, instance selection, resets, lower bound computation, counts successful and failed move attempts broken down by mode, and the sequence of item/bins for successful moves since last new instance or new mode.
