### 2021-06-10 8e1d330
Added tests for mode int to string methods
