### 2021-06-10 3cf4d5d
tests for SessionLog
