### 2021-06-11 b238d21
Added session menu and session log dialog
