### 2021-06-11 fc53132
Session log formatting
