+++ src/main/java/org/cicirello/ibp/SessionLog.java
@@ -24,6 +24,8 @@
  import java.io.Serializable;
  import java.util.ArrayList;
  import java.util.Arrays;
+ import java.util.Date;
+ import java.time.Duration;
  
  /**
  * This class is used to maintain a log of the activity of 
@@ -57,7 +59,7 @@ public final class SessionLog implements Serializable {
 		records = new RecordList();
 		addEntry("SESSION", "START");
 		currentMode = 0;
-		currentInstance = null;
+		currentInstance = "Default";
 		successfulMoves = new int[5];
 		failedMoves = new int[5];
 		currentItemSequence = new ArrayList<Item>();
@@ -137,6 +139,12 @@ public final class SessionLog implements Serializable {
 	 * @return session log formatted in html for in application viewing
 	 */
 	public String formatSessionLog() {
+		ArrayList<String> alertList = new ArrayList<String>();
+		String summary = formatSummaryStats();
+		String allActions = formatAllLoggedActions(alertList);
+		
+		String allAlerts = formatAlerts(alertList);
+		
 		StringBuilder html = new StringBuilder();
 		html.append("<html>\n<body>\n<h1><a name=\"TOP\"></a>Session Log</h1>\n<hr>\n");
 		
@@ -152,7 +160,7 @@ public final class SessionLog implements Serializable {
 		html.append("<hr>\n");
 		html.append("<h2><a name=\"summary\">Summary Statistics</h2>\n");
 		
-		// HERE SUMMARY MUST BE GENERATED
+		html.append(summary);
 		
 		html.append("<p>Return to <a href=\"#TOP\">Top</a> or <a href=\"#ToC\">Table of Contents</a>.</p>\n");
 		
@@ -160,8 +168,8 @@ public final class SessionLog implements Serializable {
 		html.append("<h2><a name=\"alerts\">Alerts</h2>\n");
 		html.append("This section is meant for instructors viewing session logs\n");
 		html.append("submitted by students for assignments. Although we believe\n");
-		html.append("that the amount of effort it would take to falsify a session\n");
-		html.append("log is greater than the effort necessary to just complete the\n");
+		html.append("that the amount of effort necessary to falsify a session\n");
+		html.append("log is greater than the effort necessary to complete the\n");
 		html.append("tutorial and any exercises assigned by the instructor,\n");
 		html.append("the application performs some rudimentary analysis to detect\n");
 		html.append("questionable records. This is where you will\n");
@@ -177,6 +185,7 @@ public final class SessionLog implements Serializable {
 		html.append("claimed by the session log.</p>\n");
 		
 		// HERE ALERTS MUST BE GENERATED
+		html.append(allAlerts);
 		
 		html.append("<p>Return to <a href=\"#TOP\">Top</a> or <a href=\"#ToC\">Table of Contents</a>.</p>\n");
 		
@@ -190,12 +199,139 @@ public final class SessionLog implements Serializable {
 		html.append("<hr>\n");
 		html.append("<h2><a name=\"logs\">All Logged Actions</h2>\n");
 		
-		// HERE ALL LOGGED ACTIONS MUST BE REPORTED
+		html.append(allActions);
 		
 		html.append("<p>Return to <a href=\"#TOP\">Top</a> or <a href=\"#ToC\">Table of Contents</a>.</p>\n");
 		
 		html.append("<hr>\n</body>\n</html>");
 		return html.toString();
+	}
+	
+	/**
+	 * Formats all of the logged actions, including with timestamps.
+	 * @param alertList A list of any alerts.
+	 * @return An html table with all logged actions.
+	 */
[…]
