### 2021-06-14 68d52b2
namespaced resources with package name
