### 2021-06-14 a6e256d
Template for session log as resource
