### 2021-06-14 59ab4f9
Session log generation
