+++ src/main/java/org/cicirello/ibp/SessionLog.java
@@ -28,6 +28,9 @@
  import java.util.Arrays;
  import java.util.Date;
  import java.util.Formatter;
+ import java.util.HashMap;
+ import java.util.HashSet;
+ import java.util.Random;
  import java.time.Duration;
  
  /**
@@ -42,7 +45,7 @@
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
  */
-public final class SessionLog implements Serializable {
+final class SessionLog implements Serializable {
 	 
 	private static final long serialVersionUID = 1L;
 	 
@@ -145,24 +148,322 @@ public final class SessionLog implements Serializable {
 		ArrayList<String> alertList = new ArrayList<String>();
 		String summary = formatSummaryStats();
 		String allActions = formatAllLoggedActions(alertList);
+		String allCompletions = formatCompletions(alertList);
 		String allAlerts = formatAlerts(alertList);
-		String allCompletions = "";
 		
+		String logString = "<html><body><h1>Session Log</h1><hr><h2>Something went wrong loading session log.</h2></body></html>";
 		try {
 			InputStream in = InteractiveBinPacking.class.getResourceAsStream("html/sessionLogTemplate.html");
 			String template = new String(in.readAllBytes());
 			in.close();
-			Formatter f = new Formatter();
-			return f.format(
+			logString = new Formatter().format(
 				template, 
 				summary, 
 				allAlerts, 
 				allCompletions, 
 				allActions
 			).toString();
-		} catch(IOException ex) {
-			return "<html><body><h1>Session Log</h1><hr><h2>Something went wrong loading session log.</h2></body></html>";
+		} catch(IOException ex) { }
+		
+		return logString;
+	}
+	
+	String formatCompletions(ArrayList<String> alertList) {
+		StringBuilder s = new StringBuilder();
+		String currentModeName = "";
+		String currentInstance = "Default";
+		ArrayList<String> completionTableRows = new ArrayList<String>();
+		int[] modeInstanceCount = new int[successfulMoves.length];
+		for (int i = 0; i < records.size(); i++) {
+			LogRecord log = records.get(i);
+			String type = log.getType();
+			if (type.equals("SET_MODE")) {
+				currentModeName = log.getData();
+			} else if (type.equals("SELECT_INSTANCE")) {
+				currentInstance = log.getData();
+			} else if (type.equals("COMPLETED")) {
+				String completedData = log.getData();
+				if (i+1 < records.size() && records.get(i+1).getType().equals("SOLUTION")) {
+					i++;
+					log = records.get(i);
+					String solutionData = log.getData();
+					int m = validateSolution(
+						completedData, 
+						solutionData, 
+						currentModeName, 
+						currentInstance, 
+						completionTableRows, 
+						alertList
+					);
+					if (m >= 0) {
+						modeInstanceCount[m]++;
+					}
+				} else {
+					alertList.add("Completed record is missing required solution.");
+				}
+			} else if (type.equals("SOLUTION")) {
+				alertList.add("Solution found without corresponding completion record.");
+			}
 		}
+		for (int i = 1; i < successfulMoves.length; i++) {
+			if (successfulMoves[i] < 20*modeInstanceCount[i]) {
+				alertList.add("Fewer moves were recorded than needed to solve the instances claimed to have been solved.");
+			}
+		}
+		if (completionTableRows.size() == 0) {
[…]
