### 2021-06-14 e14f953
Added check of number formatting
