+++ src/test/java/org/cicirello/ibp/NonGUITestCases.java
@@ -892,6 +892,211 @@ public class NonGUITestCases {
 		}
 	}
 	
+	@Test
+	public void testSessionLogFormatSummaryStats() {
+		SessionLog log = new SessionLog();
+		log.recordMove(new Item("A",22), new Bin("Bin 1", 1));
+		log.recordMove(new Item("B",22), new Bin("Bin 1", 1));
+		log.recordMove(new Item("C",22), new Bin("Bin 1", 1));
+		log.recordFailedMove();
+		log.recordMove(new Item("D",22), new Bin("Bin 1", 1));
+		log.recordMove(new Item("E",22), new Bin("Bin 1", 1));
+		
+		log.addEntry("SET_MODE", "1");
+		log.recordMove(new Item("F",22), new Bin("Bin 1", 1));
+		log.recordFailedMove();
+		log.recordMove(new Item("G",22), new Bin("Bin 1", 1));
+		log.recordMove(new Item("H",22), new Bin("Bin 1", 1));
+		log.recordFailedMove();
+		log.recordMove(new Item("I",22), new Bin("Bin 1", 1));
+		
+		log.addEntry("SET_MODE", "2");
+		log.recordFailedMove();
+		log.recordMove(new Item("J",22), new Bin("Bin 1", 1));
+		log.recordMove(new Item("K",22), new Bin("Bin 1", 1));
+		log.recordFailedMove();
+		log.recordMove(new Item("L",22), new Bin("Bin 1", 1));
+		log.recordFailedMove();
+		
+		log.addEntry("SET_MODE", "3");
+		log.recordFailedMove();
+		log.recordMove(new Item("M",22), new Bin("Bin 1", 1));
+		log.recordFailedMove();
+		log.recordFailedMove();
+		log.recordMove(new Item("N",22), new Bin("Bin 1", 1));
+		log.recordFailedMove();
+		
+		log.addEntry("SET_MODE", "4");
+		log.recordFailedMove();
+		log.recordFailedMove();
+		log.recordFailedMove();
+		log.recordMove(new Item("O",22), new Bin("Bin 1", 1));
+		log.recordFailedMove();
+		log.recordFailedMove();
+		
+		String summary = log.formatSummaryStats();
+		int i = summary.indexOf("practice");
+		int j = summary.indexOf("</tr>", i+8);
+		int x = summary.indexOf(">5</td>", i+8);
+		int y = summary.indexOf(">1</td>", i+8);
+		assertTrue(x > 0);
+		assertTrue(x < y);
+		assertTrue(y > 0);
+		
+		i = summary.indexOf("first-fit");
+		j = summary.indexOf("</tr>", i+9);
+		x = summary.indexOf(">4</td>", i+9);
+		y = summary.indexOf(">2</td>", i+9);
+		
+		i = summary.indexOf("first-fit decreasing");
+		j = summary.indexOf("</tr>", i+20);
+		x = summary.indexOf(">3</td>", i+20);
+		y = summary.indexOf(">3</td>", x+7);
+		
+		i = summary.indexOf("best-fit");
+		j = summary.indexOf("</tr>", i+8);
+		x = summary.indexOf(">2</td>", i+8);
+		y = summary.indexOf(">4</td>", i+8);
+		
+		i = summary.indexOf("best-fit decreasing");
+		j = summary.indexOf("</tr>", i+21);
+		x = summary.indexOf(">1</td>", i+8);
+		y = summary.indexOf(">5</td>", i+8);
+	}
+	
+	@Test
+	public void testSessionLogAlertDetectors() {
+		SessionLog log = new SessionLog();
+		ArrayList<String> alerts = new ArrayList<String>();
+		
+		String s = log.malformed(alerts);
+		assertTrue(s.indexOf("MALFORMED")>=0);
[…]
