### 2021-06-14 1222eb9
Test cases for SessionLog
