### 2021-06-15 f52343b
Moved SessionLog tests to dedicated class
