+++ src/test/java/org/cicirello/ibp/SessionLogTests.java
@@ -1,0 +1,376 @@
+/*
+ * Interactive Bin Packing.
+ * Copyright (C) 2008, 2010, 2020-2021  Vincent A. Cicirello
+ *
+ * This file is part of Interactive Bin Packing.
+ * 
+ * Interactive Bin Packing is free software: 
+ * you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Interactive Bin Packing is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+ 
+package org.cicirello.ibp;
+
+import org.junit.*;
+import static org.junit.Assert.*;
+import java.util.ArrayList;
+import java.io.ObjectOutputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.ObjectInputStream;
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
+
+/**
+ * JUnit tests for the SessionLog class of the
+ * Interactive Bin Packing Application.
+ */
+public class SessionLogTests {
+	
+		@Test
+	public void testSessionLogEqualsHashCodeSerialization() {
+		// Different timestamps on otherwise identical logs.
+		SessionLog log1 = new SessionLog();
+		try {
+			Thread.sleep(10);
+		} catch (InterruptedException ex) {}
+		SessionLog log2 = new SessionLog();
+		assertNotEquals(log1, log2);
+		
+		try {
+			// Equal case
+			ByteArrayOutputStream outBytes = new ByteArrayOutputStream();
+			ObjectOutputStream out = new ObjectOutputStream(outBytes);
+			out.writeObject(log1);
+			out.flush();
+			out.close();
+			byte[] bytes = outBytes.toByteArray();
+			ByteArrayInputStream inBytes = new ByteArrayInputStream(bytes);
+			ObjectInputStream in = new ObjectInputStream(inBytes);
+			SessionLog deserialized = (SessionLog)in.readObject();
+			in.close();
+			assertEquals(log1, deserialized);
+			assertEquals(log1.hashCode(), deserialized.hashCode());
+			
+			// same type different data
+			log1.addEntry("T1", "D1");
+			deserialized.addEntry("T1", "D2");
+			assertNotEquals(log1, deserialized);
+			
+			// multiple records equal case
+			outBytes = new ByteArrayOutputStream();
+			out = new ObjectOutputStream(outBytes);
+			out.writeObject(log1);
+			out.flush();
+			out.close();
+			bytes = outBytes.toByteArray();
+			inBytes = new ByteArrayInputStream(bytes);
+			in = new ObjectInputStream(inBytes);
+			deserialized = (SessionLog)in.readObject();
+			in.close();
+			assertEquals(log1, deserialized);
+			assertEquals(log1.hashCode(), deserialized.hashCode());
+			
+			// different num records case
+			log1.addEntry("T1", "D2");
+			assertNotEquals(log1, deserialized);
[…]
