### 2021-06-15 d65ff56
removed unnecessary check
