### 2021-06-15 d0c8acc
test cases checkBins
