+++ src/test/java/org/cicirello/ibp/SessionLogTests.java
@@ -127,7 +127,7 @@ public class SessionLogTests {
 	}
 	
 	@Test
-	public void testSessionLogFormatSummaryStats() {
+	public void testFormatSummaryStats() {
 		SessionLog log = new SessionLog();
 		log.recordMove(new Item("A",22), new Bin("Bin 1", 1));
 		log.recordMove(new Item("B",22), new Bin("Bin 1", 1));
@@ -199,7 +199,7 @@ public class SessionLogTests {
 	}
 	
 	@Test
-	public void testSessionLogAlertDetectors() {
+	public void testAlertDetectors() {
 		SessionLog log = new SessionLog();
 		ArrayList<String> alerts = new ArrayList<String>();
 		
@@ -251,7 +251,7 @@ public class SessionLogTests {
 	}
 	
 	@Test
-	public void testSessionLogFormatActions() {
+	public void testFormatActions() {
 		SessionLog log = new SessionLog();
 		ArrayList<String> alerts = new ArrayList<String>();
 		
@@ -286,7 +286,7 @@ public class SessionLogTests {
 	}
 	
 	@Test
-	public void testSessionLogSolutionItemOrdering() {
+	public void testSolutionItemOrdering() {
 		SessionLog log = new SessionLog();
 		ArrayList<String> alerts = new ArrayList<String>();
 		
@@ -319,7 +319,7 @@ public class SessionLogTests {
 	}
 	
 	@Test
-	public void testSessionLogCheckBinOrdering() {
+	public void testSolutionBinOrdering() {
 		int[] sizesSorted = { 50, 35, 35, 30, 25, 10, 8, 8, 5 };
 		int[] ffBins =      {  1,  1,  2,  2,  2,  1, 2, 3, 1 };
 		int[] bfBins =      {  1,  1,  2,  2,  2,  2, 1, 3, 1 };
@@ -363,7 +363,103 @@ public class SessionLogTests {
 	}
 	
 	@Test
-	public void testSessionLogExtractMethods() {
+	public void testBinCount() {
+		SessionLog log = new SessionLog();
+		
+		int[][] cases = {
+			{},
+			{1, 1, 1, 1, 1, 1},
+			{1, 2, 2, 1, 2, 1},
+			{3, 2, 3, 1, 2, 3},
+			{3, 2, 4, 1, 4, 3},
+			{5, 2, 4, 1, 4, 3},
+			{5, 2, 4, 1, 6, 3}
+		};
+		for (int i = 0; i < cases.length; i++) {
+			assertEquals(i, log.binCount(cases[i]));
+		}
+	}
+	
+	@Test
+	public void testCheckInstance() {
+		SessionLog log = new SessionLog();
+		ArrayList<String> alerts = new ArrayList<String>();
+		
+		int[] defaultSizes = {
+			36, 33, 39, 43, 7, 19, 37, 8, 29, 28, 37, 23, 29, 10, 22, 11, 33, 9, 17, 30
+		};
+		
+		int[] defaultWrong = {
+			37, 33, 39, 43, 7, 19, 37, 8, 29, 29, 37, 23, 29, 10, 22, 11, 33, 9, 17, 31
+		};
+		
+		String[] items = {
+			"A", "B", "C", "D", "E", "F", "G", "H", "I", "J", 
+			"K", "L", "M", "N", "O", "P", "Q", "R", "S", "T"
+		};
+		
+		String[] duplicateItems = {
+			"A", "B", "C", "D", "E", "F", "G", "H", "I", "J", 
+			"K", "L", "M", "N", "O", "P", "Q", "R", "S", "A"
+		};
+		
+		String[] unknownItems = {
[…]
