### 2021-06-15 8c31e54
test case validateSolution
