### 2021-06-15 4433999
test cases validateSolution
