### 2021-06-15 83728f7
fix bug in checkBins
