### 2021-06-15 8ea2fbd
Update SessionLogTests.java
