+++ src/test/java/org/cicirello/ibp/SessionLogTests.java
@@ -570,7 +570,13 @@ public class SessionLogTests {
 			
 			"ItemSequence=Q 50 B 48 E 47 H 46 M 44 L 38 R 36 D 35 I 35 F 34 O 32 A 27 C 27 G 26 K 25 N 25 T 25 J 23 S 21 P 20, BinSequence=1 1 2 2 3 3 4 4 5 5 6 4 5 6 6 7 7 7 7 eight",
 			
-			"ItemSequence=Q 50 B 48 E 47 H 46 M 44 L 38 R 36 D 35 I 35 F 34 O 32 A 27 C 27 G 26 K 25 N 25 T 25 J 23 S 21, BinSequence=1 1 2 2 3 3 4 4 5 5 6 4 5 6 6 7 7 7 7 8"
+			"ItemSequence=Q 50 B 48 E 47 H 46 M 44 L 38 R 36 D 35 I 35 F 34 O 32 A 27 C 27 G 26 K 25 N 25 T 25 J 23 S 21, BinSequence=1 1 2 2 3 3 4 4 5 5 6 4 5 6 6 7 7 7 7 8",
+			
+			"ItemSequence=Q 50 B 48 E 47 H 46 M 44 L 38 R 36 D 35 I 35 F 34 O 32 A 27 C 27 G 26 K 25 N 25 T 25 J 23 S 21 P twenty, BinSequence=1 1 2 2 3 3 4 4 5 5 6 4 5 6 6 7 7 7 7 8",
+			
+			"ItemSequence=Q 50 B 48 E 47 H 46 M 44 L 38 R 36 D 35 I 35 F 34 O 32 A 27 C 27 G 26 K 25 N 25 T 25 J 23 S 21 P 20, BinSequence=1 1 2 2 3 3 4 4 5 5 6 4 5 6 6 7 7 7 8 8",
+			
+			"ItemSequence=Q 50 B 48 E 47 H 46 M 44 L 38 R 36 D 35 I 35 F 34 O 32 A 27 C 27 G 26 K 25 N 25 T 25 J 23 P 20 S 21, BinSequence=1 1 2 2 3 3 4 4 5 5 6 4 5 6 6 7 7 7 7 8"
 		};
 		
 		for (int i = 0; i < badSolutions.length; i++) {
@@ -582,9 +588,109 @@ public class SessionLogTests {
 			assertEquals(5+i, alertList.size());
 			assertEquals(4, completionTableRows.size());
 		}
+		
+		String badRandomSolution = "ItemSequence=Q 50 B 48 E 47 H 46 M 44 L 38 R 36 D 35 I 35 F 34 O 32 A 27 C 27 G 26 K 25 N 25 T 25 J 23 S 21 P 19, BinSequence=1 1 2 2 3 3 4 4 5 5 6 4 5 6 6 7 7 7 7 8";
+		
+		assertEquals(-1, log.validateSolution(
+			String.format(completedTemplate, modeNum, "Random", modeName),
+			badRandomSolution, modeName, "Random",
+			completionTableRows, alertList
+		));
+		int numAlerts = 5+badSolutions.length;
+		assertEquals(numAlerts, alertList.size());
+		assertEquals(4, completionTableRows.size());
 	}
 	
+	@Test
+	public void testFormatCompletions() {
+		SessionLog log = new SessionLog();
+		ArrayList<String> alertList = new ArrayList<String>();
+		int modeNum = 1;
+		String modeName = "first-fit";
+		String selectInstance = "#" + Integer.MAX_VALUE;
+		String completedTemplate = "ModeNum=%d, Instance=%s, Mode=%s";
+		
+		log.addEntry("SET_MODE", ""+modeNum);
[…]
