### 2021-06-15 a8d4f5a
Update SessionLogTests.java
