### 2021-06-17 c44de6a
Added logo to help, tutorial, about dialogs
