### 2021-06-17 1f916d4
Formatting session log file contents
#37
