### 2021-06-17 b72ee7e
save session log to file
#37
