+++ src/main/java/org/cicirello/ibp/SessionLog.java
@@ -19,19 +19,19 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
  
- package org.cicirello.ibp;
+package org.cicirello.ibp;
  
- import java.io.Serializable;
- import java.io.InputStream;
- import java.io.IOException;
- import java.util.ArrayList;
- import java.util.Arrays;
- import java.util.Date;
- import java.util.Formatter;
- import java.util.HashMap;
- import java.util.HashSet;
- import java.util.Random;
- import java.time.Duration;
+import java.io.Serializable;
+import java.io.InputStream;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Date;
+import java.util.Formatter;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Random;
+import java.time.Duration;
  
  /**
  * This class is used to maintain a log of the activity of 
