### 2021-06-18 f58c25b
save session log dialog
#37
