### 2021-06-18 3c68bb3
load session logs
