### 2021-06-21 a1e64fc
tests for SessionLog save and load
#37
