### 2021-06-21 fae73d8
test cases for SessionLog read and write
