+++ src/test/java/org/cicirello/ibp/SessionLogTests.java
@@ -40,15 +40,86 @@ import java.io.PrintWriter;
 public class SessionLogTests {
 	
 	@Test
-	public void testWriteReadFiles() {
+	public void testWriteReadFilesFromApplicationState() {
 		ApplicationState state = createApplicationState();
 		StringWriter sOut = writeSessionLogToString(state);
 		SessionLog savedLog = readSessionLogFromString(sOut);
 		assertTrue(state.equalsInternalSessionLog(savedLog));
+		
 		sOut = writeSessionLogToString(state);
-		String asString = readSessionLogAsStringFromString(sOut, state); 
-		int index = asString.indexOf("SAVE_SESSION_LOG");
-		assertTrue(index>=0);
+		String asString = readSessionLogAsStringFromString(sOut, state, false); 
+		assertTrue(asString.indexOf("SAVE_SESSION_LOG")>=0);
+		
+		sOut = writeSessionLogToString(state);
+		assertNull(readSessionLogAsStringFromString(sOut, state, true));
+	}
+	
+	@Test
+	public void testWriteReadFilesFromSessionLog() {
+		SessionLog log = new SessionLog();
+		for (int m = 0; m < 5; m++) {
+			log.addEntry("SET_MODE", ""+m);
+			for (int i = 0; i < 20; i++) {
+				String n = ""+('A'+i);
+				log.recordMove(new Item(n, 20), new Bin("Bin 1", 1));
+			}
+			for (int i = 0; i < m; i++) {
+				log.recordFailedMove();
+			}
+		}
+		StringWriter sOut = writeSessionLogToString(log);
+		SessionLog savedLog = readSessionLogFromString(sOut);
+		assertEquals(log, savedLog);
+		String asString = savedLog.toString();
+		
+		// break asString in various ways
+		assertNull(readSessionLogFromString(asString.replace("<session>","<sesion>")));
+		assertNull(readSessionLogFromString(asString.replace("</session>","</sesion>")));
+		assertNull(readSessionLogFromString(asString.replace("<moveCounts>","<moves>")));
+		assertNull(readSessionLogFromString(asString.replace("</moveCounts>","</moves>")));
+		assertNull(readSessionLogFromString(asString.replace("<actions>","<actons>")));
+		assertNull(readSessionLogFromString(asString.replace("</actions>","</actons>")));
+		assertNull(readSessionLogFromString(asString.replace("<action>","<acton>")));
+		assertNull(readSessionLogFromString(asString.replace("</action>","</acton>")));
+		assertNull(readSessionLogFromString(asString.replace("<type>","<typ>")));
+		assertNull(readSessionLogFromString(asString.replace("</type>","</typ>")));
+		assertNull(readSessionLogFromString(asString.replace("<data>","<dat>")));
+		assertNull(readSessionLogFromString(asString.replace("</data>","</dat>")));
+		assertNull(readSessionLogFromString(asString.replace("<timestamp>","<timestmp>")));
+		assertNull(readSessionLogFromString(asString.replace("</timestamp>","</timestmp>")));
+		assertNull(readSessionLogFromString(asString.replace("<successful>","<sucessful>")));
+		assertNull(readSessionLogFromString(asString.replace("</successful>","</sucessful>")));
+		assertNull(readSessionLogFromString(asString.replace("<failed>","<fail>")));
+		assertNull(readSessionLogFromString(asString.replace("</failed>","</fail>")));
+		
+		// break asString with extra stuff
+		assertNull(readSessionLogFromString(asString + "extra"));
+		
+		// break asString by removing required stuff
+		asString = asString.substring(0, asString.indexOf("</session>"));
+		assertNull(readSessionLogFromString(asString));
[…]
