### 2021-06-21 abb95e4
test cases for session log loading
#37
