### 2021-06-21 71c014e
test cases save and load session log
#37
