### 2021-06-21 4fa87c1
removed check for impossible condition
