### 2021-06-22 3b72750
minor optimizations
