### 2021-06-23 38c6dd9
test cases for save and load session log
#37
