+++ src/test/java/org/cicirello/ibp/GUITestCases.java
@@ -23,6 +23,8 @@ package org.cicirello.ibp;
 
 import org.junit.*;
 import static org.junit.Assert.*;
+import org.junit.rules.TemporaryFolder;
+
 import java.util.ArrayList;
 import javax.swing.JMenu;
 import java.awt.BorderLayout;
@@ -35,6 +37,10 @@ import javax.swing.JComboBox;
 import javax.swing.JButton;
 import java.awt.GraphicsEnvironment;
 import java.io.File;
+import java.io.FileReader;
+import java.nio.charset.StandardCharsets;
+import javax.swing.JOptionPane;
+import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
 
@@ -50,6 +56,9 @@ import java.net.URISyntaxException;
  *
  */
 public class GUITestCases {
+	
+	@Rule
+    public TemporaryFolder tempFolder = new TemporaryFolder();
 	
 	@Test
 	public void testInfoMenu() {
@@ -339,6 +348,119 @@ public class GUITestCases {
 		assertFalse(((MenuBarTester)menus).lowerBoundButtonClicked);
 		opsMenu.getItem(2).doClick();
 		assertTrue(((MenuBarTester)menus).lowerBoundButtonClicked);
+	}
+	
+	@Test
+	public void testSessionMenuSaveSession() {
+		class MenuBarTester extends MenuBar {
+			int errorCount;
+			int confirmCount;
+			int confirmSaveResponse;
+			
+			MenuBarTester(InteractiveBinPacking f, ApplicationState state) {
+				super(f, state);
+				errorCount = confirmCount = 0;
+			}
+			
+			@Override
+			void displayErrorMessage(String message) {
+				errorCount++;
+			}
+			
+			@Override
+			int confirmSave() {
+				confirmCount++;
+				return confirmSaveResponse;
+			}
+			
+			void setResponse(int response) {
+				confirmSaveResponse = response;
+			}
+		}
+		int[] sizes = { 7, 2, 18, 3, 6 };
+		CallBack cb = new CallBack() {
+			@Override public void call() { }
+		};
+		Floor floor = new Floor(sizes);
+		ApplicationState state = new ApplicationState(1, floor, cb, cb, cb);
+		MenuBarTester menus = new MenuBarTester(null, state);
+		
+		try {
+			File sub = tempFolder.newFolder("sub");
+			File logFile = new File(sub, "testlog.ibp");
+			menus.saveSessionLog(logFile);
+			int index = 0;
+			try (FileReader in = new FileReader(logFile, StandardCharsets.UTF_8)) {
+				String session = state.loadSessionLog(in);
+				index = session.indexOf("SAVE_SESSION_LOG");
+				assertTrue(index>=0);
+			} catch(IOException ex) {
+				fail();
+			}
+			assertEquals(0, menus.errorCount);
+			assertEquals(0, menus.confirmCount);
+			
+			File without = new File(sub, "without.txt");
+			File with = new File(sub, "without.txt.ibp");
+			menus.saveSessionLog(without);
+			try (FileReader in = new FileReader(with, StandardCharsets.UTF_8)) {
+				String session = state.loadSessionLog(in);
+				index++;
+				index = session.indexOf("LOAD_SESSION_LOG", index);
+				assertTrue(index>=0);
+				index++;
+				index = session.indexOf("SAVE_SESSION_LOG", index);
+				assertTrue(index>=0);
+			} catch(IOException ex) {
+				fail();
+			}
+			assertEquals(0, menus.errorCount);
+			assertEquals(0, menus.confirmCount);
+			
[…]
