### 2021-06-23 173e61c
easy validation version of assignment 1
#38
