+++ assignments/2/assignment.tex
@@ -1,0 +1,406 @@
+% Copyright 2021 Vincent A. Cicirello
+%
+% License: CC BY-NC-SA (Creative Commons Attribution-NonCommercial-ShareAlike)
+% Full license text link: https://creativecommons.org/licenses/by-nc-sa/4.0/
+% Short summary of license: Allows remixing and adapting non-commercially, 
+% requires attribution, and adaptations required to be licensed under 
+% identical terms.
+%
+% Original version of this assignment can be found in the assignments directory
+% in this GitHub repository: https://github.com/cicirello/InteractiveBinPacking
+%
+% Instructors: If you are an instructor of a course using this assignment
+% keep an eye out for comments addressed to you throughout this file.
+
+\documentclass[11pt,letterpaper]{article}
+
+\usepackage{times}
+\usepackage{url}
+\usepackage{graphicx}
+\usepackage{hyperref}
+\usepackage{enumitem}
+
+\setlength{\textwidth}{6.5in}
+\setlength{\oddsidemargin}{0in}
+\setlength{\evensidemargin}{0in}
+\setlength{\topmargin}{0in}
+\setlength{\headheight}{0in}
+\setlength{\headsep}{0in}
+\setlength{\textheight}{9in}
+
+\begin{document}
+
+\title{Combinatorial Optimization, the Bin Packing Problem, 
+and Constructive Heuristics}
+
+% Instructors: You might use the author field for the name
+% of the course or perhaps the assignment number. I've
+% left the author blank for now.
+\author{}
+
+% Instructors: You might insert the date of the assignment
+% here. I've left the date blank for now.
+\date{}
+
+\maketitle
+
+\section{Before We Begin}\label{sec:before}
+
+% Instructors: You might need to customize this section with your
+% own special instructions, such as how to submit the assignment,
+% deadline, etc. Especially note the paragraph below that indicates
+% using student id for the instance number later in the assignment.
+
+\paragraph*{Problem Instance Numbers:}
+Later in this assignment, you will work through a few exercises,
+and each student's exercises will be different. Specifically, 
+each student will use a different instance of the bin packing problem.
+When you get to the exercises in Sections~\ref{sec:ff},~\ref{sec:ffd},~\ref{sec:bf}, 
+and~\ref{sec:bfd}, you will be asked to enter your problem instance number.
+This will be uniquely determined for each student by your student id number.
+It must be an integer, so start by dropping any letters or other non-numeric
+characters from your student id number. If there are any 0s at the start, then
+drop those as well. What remains is your problem instance number.
+For example, if your student id number is ABC00012345678900, then your
+problem instance number is 12345678900. In the application that you are using
+the problem instance numbers are values of type long (in Java).
+
+\paragraph*{Session Logs:}
+In the application that we will use, under the ``Session'' menu, you will
+find options to save a session log. This will generate a file with an 
+extension of ``*.ibp''. You can otherwise name the file anything you'd like.
+If you work on the assignment over multiple sessions, just save each of the 
+sessions. Make sure you don't select the same file as your previous session,
+or else it will be overwritten. Don't attempt to edit the session log files
+manually. Doing so will very likely generate alerts when your instructor
+views the session logs, possibly invalidating your work.
+
+The session log files are what you will be submitting for this assignment,
+in addition to answers to any questions found later in this assignment. If you
+spread your work over multiple sessions, then just submit the log files for all
+of your sessions.
+
+\paragraph*{What Your Instructor Will be Looking for in the Session Logs:}
+The instructor of your course, in addition to the answers to questions later in this
+document, will expect to find a record of the following in the session logs that
+you submit: (1) that you completed the ``Default'' problem instance in all 4
[…]
