+++ assignments/2/assignment.tex
@@ -402,5 +402,5 @@ computation)? If yes, explain why. If no, explain why we don't know for sure.
 
 \vspace*{0.5in}
 
-
+\textbf{\large IMPORTANT: Make sure you save your session log to submit with your assignment.}
 \end{document}
