+++ README.md
@@ -59,6 +59,15 @@ or self-learners) work through the
 
 The prebuilt jar of the application is built with the OpenJDK 11.
 
+## Example Assignments
+
+In addition to the Interactive Bin Packing application, the
+repository contains a directory of [Example Assignments](assignments)
+that utilize the application. Self-directed learners may begin with
+[assignments/1](assignments/1); while course instructors may prefer
+to assign their students [assignments/2](assignments/2), which utilizes
+the Session Log functionality introduced in version 3.1.0 to make it easy
+for instructors to validate students' work. 
 
 ## User Documentation
 
