+++ README.md
@@ -1,8 +1,9 @@
 # Interactive Bin Packing
 
-Copyright (C) 2008, 2010, 2020-2021 Vincent A. Cicirello.
+[<img alt="Interactive Bin Packing" 
+     src="images/ibp.png" width="640">](#interactive-bin-packing)
 
-https://www.cicirello.org/
+Copyright (C) 2008, 2010, 2020-2021 [Vincent A. Cicirello](https://www.cicirello.org/).
 
 | | |
 | :--- | :--- |
