### 2021-07-13 #87 @labarba
@adeepak7  – hi! 👋
A while back you signed up to be a reviewer for JOSE, the Journal of Open Source Education. I wonder if you could help us out with this submission here:
**Interactive Bin Packing: A Java Application for Learning Constructive Heuristics for Combinatorial Optimization**
[View article proof on GitHub](https://github.com/openjournals/jose-papers/blob/jose.00087/jose.00087/10.21105.jose.00087.pdf)
It's a Java application that generates random instances of the bin packing problem, to be used as a tutorial and an activity for students.
