### 2021-07-14 #87 @adeepak7
@labarba Thank you for giving me this opportunity. Having worked on standalone JAVA applications and their respective deployments, I promise to provide constructive feedback and raise relevant issues.
