+++ paper/paper.bib
@@ -1,106 +1,0 @@
-@book{Garey1979,
-  author    = {Michael R. Garey and
-               David S. Johnson},
-  title     = {Computers and Intractability: {A} Guide to the Theory of NP-Completeness},
-  publisher = {W. H. Freeman},
-  year      = {1979},
-  isbn = {0716710455}
-}
-
-@INPROCEEDINGS{Leinberger1999,
-author = {W. Leinberger and G. Karypis and V. Kumar},
-booktitle = {Proceedings of the 1999 International Conference on Parallel Processing},
-title = {Multi-Capacity Bin Packing Algorithms with Applications to Job Scheduling under Multiple Constraints},
-year = {1999},
-pages = {404--412},
-doi = {10.1109/ICPP.1999.797428},
-publisher = {IEEE Computer Society}
-}
-
-@article{Malkevitch2004b,
-  title={Bin packing and machine scheduling},
-  author={Malkevitch, Joseph},
-  journal={Feature column from the AMS: Monthly essays on mathematical topics},
-  year={2004},
-  month = {June}
-}
-
-@article{Malkevitch2004a,
-  title={Bin packing},
-  author={Malkevitch, Joseph},
-  journal={Feature column from the AMS: Monthly essays on mathematical topics},
-  year={2004},
-  month = {May}
-}
-
-@article{Wang2006,
-author = {Wang, Chonggang and Li, Bo and Sivalingam, Krishna and Sohraby, Kazem},
-title = {Scalable Multiple Channel Scheduling with Optimal Utility in Wireless Local Area Networks},
-year = {2006},
-publisher = {Springer-Verlag},
-volume = {12},
-number = {2},
-doi = {10.1007/s11276-005-5266-y},
-journal = {Wireless Networks},
-pages = {189--198}
-}
-
-@inproceedings{Nishida2009,
-author = {Nishida, Tomohiro and Kanemune, Susumu and Idosaka, Yukio and Namiki, Mitaro and Bell, Tim and Kuno, Yasushi},
-title = {A CS Unplugged Design Pattern},
-year = {2009},
-publisher = {ACM},
-doi = {10.1145/1508865.1508951},
-booktitle = {Proceedings of the 40th ACM Technical Symposium on Computer Science Education},
-pages = {231--235}
-}
-
-@inproceedings{Levitin2005,
-author = {Levitin, Anany},
-title = {Analyze That: Puzzles and Analysis of Algorithms},
-year = {2005},
-publisher = {ACM},
-doi = {10.1145/1047344.1047409},
-booktitle = {Proceedings of the 36th SIGCSE Technical Symposium on Computer Science Education},
-pages = {171--175}
-}
-
-@article{Cicirello2013,
-author = {Cicirello, Vincent A.},
-title = {A CS Unplugged Activity for the Online Classroom},
-year = {2013},
-volume = {28},
-number = {6},
-journal = {Journal of Computing Sciences in Colleges},
-month = {June},
-pages = {162--168}
-}
-
-@article{Cicirello2009,
-author = {Cicirello, Vincent A.},
-title = {Collective Bin Packing: An Active Learning Exercise},
-year = {2009},
-volume = {24},
-number = {6},
-journal = {Journal of Computing Sciences in Colleges},
-month = {June},
-pages = {117--123}
-}
-
-@article{Cicirello2007,
-author = {Cicirello, Vincent A.},
-title = {An Interdisciplinary Course on Artificial Intelligence Designed for a Liberal Arts Curriculum},
-year = {2008},
-volume = {23},
-number = {3},
-journal = {Journal of Computing Sciences in Colleges},
-month = {January},
-pages = {120–127}
-}
-
-
-
-
-
-
-
