+++ paper/paper.md
@@ -1,141 +1,0 @@
----
-title: 'Interactive Bin Packing: A Java Application for Learning Constructive Heuristics for Combinatorial Optimization'
-tags:
-  - combinatorial optimization
-  - bin packing
-  - constructive heuristics
-  - discrete optimization
-  - Java
-  - self-guided tutorial
-authors:
-  - name: Vincent A. Cicirello
-    orcid: 0000-0003-1072-8559
-    affiliation: 1
-affiliations:
-  - name: Computer Science, School of Business, Stockton University, Galloway, NJ 08205
-    index: 1
-date: 3 July 2020
-bibliography: paper.bib
----
-
-# Summary
-
-Interactive Bin Packing provides a self-guided tutorial on combinatorial 
-optimization, the bin packing problem, and constructive heuristics. It also 
-enables users to interact with bin packing instances to explore their 
-own problem solving strategies, or to test their knowledge of the 
-constructive heuristics covered by the tutorial. The application is not a 
-solver for bin packing, but rather it is a tool for learning about the 
-bin packing problem, and for learning about heuristic techniques 
-for solving instances of the problem.
-
-The repository (https://github.com/cicirello/InteractiveBinPacking) contains 
-the source code, documentation, and example assignments suitable for courses
-on discrete mathematics, algorithms, or artificial intelligence (AI), or by 
-self-directed learners. An executable jar of the application is available 
-from GitHub Releases and Maven Central.
-
-# Bin Packing
-
-Bin packing is a combinatorial optimization problem involving 
-packing a set of items into a set of bins to minimize the number of 
-bins. It has many practical applications, including 
-within logistics, such as packing shipments into trucks or airplanes, 
-as well as applications such as creating compilations of musical works, 
-scheduling television advertisements, placing files in fixed-length 
-memory blocks on a file system, wireless network channel scheduling, 
-job scheduling, among many 
-others [@Leinberger1999; @Malkevitch2004a; @Malkevitch2004b; @Wang2006].
-
-As an NP-Hard [@Garey1979] optimization problem, it is unlikely that any 
-polynomial-time algorithm exists guaranteed to optimally solve bin packing. 
-Thus, heuristics and metaheuristics are often used. Among the topics 
-covered by this tutorial application are constructive heuristics. A 
-heuristic is a practical problem solving approach that provides a 
-solution of sufficient quality, while using relatively little time. A 
-constructive heuristic iteratively applies some rule to build upon an 
-initially empty solution until it is complete. 
-
-# Origin of the Application
-
-In my teaching, I have used bin packing to introduce heuristics to 
-students in courses for computer science majors, as well as for 
-non-majors. Using the Computer Science Unplugged [@Nishida2009] 
-methodology, I previously developed an in-class activity called 
-Collective Bin Packing [@Cicirello2009], in which students take turns 
-making individual decisions on which item to place in which bin. One 
-objective of that activity is to illustrate concepts of collective
-intelligence, or swarm intelligence, where intelligent behavior emerges 
-from indirect interaction within the swarm. A preliminary version of 
-Interactive Bin Packing was used by the instructor to keep track of the 
-state of the emerging solution. Without the aid of the application, 
-the Collective Bin Packing activity was later adapted to an online 
-course environment [@Cicirello2013], and even used in a course for 
-non-majors [@Cicirello2007]. 
-
-# Functionality
-
-This first public release of Interactive Bin Packing features a 
-self-guided tutorial that explains key concepts of combinatorial 
-optimization, bin packing, and constructive heuristics. It also 
-includes several heuristic modes, in which the user can test their 
-knowledge of the different constructive heuristics. When used in a 
-heuristic mode, the user is expected to make decisions that match 
-their chosen heuristic's behavior. The application provides feedback 
-on correctness throughout.
-
[…]
