+++ .github/workflows/build.yml
@@ -5,6 +5,7 @@ on:
     branches: [ master ]
   pull_request:
     branches: [ master ]
+  workflow_dispatch:
 
 jobs:
   build:
@@ -12,7 +13,14 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-    - uses: actions/checkout@v2
+    - name: Checkout
+      uses: actions/checkout@v2
+
+    - name: Checkout badges branch
+      uses: actions/checkout@v2
+      with:
+        ref: badges
+        path: badges
 
     - name: Set up JDK 1.11
       uses: actions/setup-java@v2
@@ -27,6 +35,7 @@ jobs:
       id: jacoco
       uses: cicirello/jacoco-badge-generator@v2
       with:
+        badges-directory: badges
         generate-branches-badge: true
 
     - name: Log coverage percentage
@@ -40,19 +49,31 @@ jobs:
         name: jacoco-report
         path: target/site/jacoco/
 
-    - name: Create pull request
-      if: ${{ github.event_name == 'push' }}
-      uses: peter-evans/create-pull-request@v3.10.0
-      with:
-        title: "Autogenerated Jacoco Coverage Badge"
-        body: > 
-          Autogenerated Jacoco coverage badge, generated by 
-          the [jacoco-badge-generator](https://github.com/cicirello/jacoco-badge-generator) 
-          GitHub action.
-        commit-message: "Autogenerated Jacoco Coverage Badge"
-        author: Vincent A Cicirello <cicirello@users.noreply.github.com>
-        committer: Vincent A Cicirello <cicirello@users.noreply.github.com>
-        branch: create-pull-request/badge
-        delete-branch: true
-        labels: |
-          automated
+    - name: Commit and push
+      if: ${{ github.event_name != 'pull_request' }}
+      run: |
+        cd badges
+        if [[ `git status --porcelain *.svg` ]]; then
+          git config --global user.name 'Vincent A. Cicirello'
+          git config --global user.email 'cicirello@users.noreply.github.com'
+          git add *.svg
+          git commit -m "Autogenerated JaCoCo coverage badges" *.svg
+          git push
+        fi
+
+#    - name: Create pull request
+#      if: ${{ github.event_name == 'push' }}
+#      uses: peter-evans/create-pull-request@v3.10.0
+#      with:
+#        title: "Autogenerated Jacoco Coverage Badge"
+#        body: > 
+#          Autogenerated Jacoco coverage badge, generated by 
+#          the [jacoco-badge-generator](https://github.com/cicirello/jacoco-badge-generator) 
+#          GitHub action.
+#        commit-message: "Autogenerated Jacoco Coverage Badge"
+#        author: Vincent A Cicirello <cicirello@users.noreply.github.com>
+#        committer: Vincent A Cicirello <cicirello@users.noreply.github.com>
+#        branch: create-pull-request/badge
+#        delete-branch: true
+#        labels: |
+#          automated
