### 2021-08-18 54ed56b
moved autgenerated badges to badges branch
