### 2021-09-16 #87 @labarba
Looks like we're short of a reviewer for this submission. Sorry for the delay, @cicirello !

@mattforshaw — You signed to review for JOSE (The Journal of Open Source Education) a while back. Would you be interested in contributing a review for this one?
**Interactive Bin Packing: A Java Application for Learning Constructive Heuristics for Combinatorial Optimization**
[View article proof on GitHub](https://github.com/openjournals/jose-papers/blob/jose.00087/jose.00087/10.21105.jose.00087.pdf)
It's a Java application that generates random instances of the bin packing problem, to be used as a tutorial and an activity for students.
