### 2021-10-05 #87 @labarba
Thank you @adeepak7 and @turketwh for agreeing to review this JOSE submission! I will now start the review, which will happen in a new GitHub issue. You will have a review checklist there. Be sure to accept the invitation you will receive from GitHub, which will open your permissions to edit on the checklist. 🙏
