### 2021-10-05 #140 @whedon
[REVIEW]: Interactive Bin Packing: A Java Application for Learning Constructive Heuristics for Combinatorial Optimization
**Submitting author:** @cicirello (<a href="http://orcid.org/0000-0003-1072-8559">Vincent A Cicirello</a>)
**Repository:** <a href="https://github.com/cicirello/InteractiveBinPacking" target ="_blank">https://github.com/cicirello/InteractiveBinPacking</a>
**Version:** v3.1.1

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode".
