### 2021-10-16 #140 @labarba
Hi  @adeepak7, @turketwh — do you have any questions about how the JOSE review process works? Feel free to ask me here.

@adeepak7 — I see you've not gotten a start yet. If you could pop in here and check off the first few items in your checklist, that'll tell me you're still with us on this! 🙏
