### 2021-11-04 cicirello/InteractiveBinPacking#51 @turketwh
Feature Req: Additions to README.md documentation related to support, installation
As I work through reviewing this application as part of a JOSE submission (the review issue thread is linked [here](https://github.com/openjournals/jose-reviews/issues/140) for reference), two initial additions to the documentation would be useful:
 - Explicit directions for requesting support with respect to the application (listed as part of the *Documentation, Community guidelines* expectations).  A particular email address and desired subject line are an example, though whatever works for you, as long as it is explicitly stated, should be OK.
 - In the *Building and Testing the Application (with Maven)* section of the `README.md` file, a link to the Maven download site? This is not a significant concern, but could be useful to users who have not used Maven before.
