### 2021-11-04 cicirello/InteractiveBinPacking#51 @turketwh
The proposed changes in the pull request satisfy the concerns expressed in my issue.
