### 2021-11-26 cicirello/InteractiveBinPacking#54 @turketwh
BUG: Missing API documentation for SessionLog class?
As I work through reviewing this application as part of a JOSE submission (the review issue thread is linked [here](https://github.com/openjournals/jose-reviews/issues/140) for reference), I think (though am not 100% sure) there may be a missing component of the Javadocs for the application's Java classes.

**Describe the bug**
No Javadoc documentation for the `SessionLog.java` class can be found in the `apidocs` subfolders after the application is built.

**To Reproduce**
Steps to reproduce the behavior:
1. Run `mvn package`
2. Explore files generated in the subfolders of the `target\apidocs` folder.

**Expected behavior**
It is unclear to me whether documentation for `SessionLog` should be generated? There is a lot of Javadoc-style information added to the `SessionLog` class file, but it has some unique features in how it is used - being defined as a `final` class and being used as the type for a `final` variable within `ApplicationState.java`.  If its usage should lead to no documentation being generated, then this issue can be closed.

**Relevant System Info (please complete the following information):**
 - Windows 10
 - javac 16.0.1
 - maven 3.8.3
