+++ .github/workflows/build.yml
@@ -22,7 +22,7 @@ jobs:
         ref: badges
         path: badges
 
-    - name: Set up JDK 1.11
+    - name: Set up JDK 11
       uses: actions/setup-java@v2
       with:
         distribution: 'adopt'
@@ -37,6 +37,7 @@ jobs:
       with:
         badges-directory: badges
         generate-branches-badge: true
+        generate-summary: true
 
     - name: Log coverage percentage
       run: |
@@ -53,27 +54,22 @@ jobs:
       if: ${{ github.event_name != 'pull_request' }}
       run: |
         cd badges
-        if [[ `git status --porcelain *.svg` ]]; then
-          git config --global user.name 'Vincent A. Cicirello'
-          git config --global user.email 'cicirello@users.noreply.github.com'
-          git add *.svg
-          git commit -m "Autogenerated JaCoCo coverage badges" *.svg
+        if [[ `git status --porcelain *.svg *.json` ]]; then
+          git config --global user.name 'github-actions'
+          git config --global user.email '41898282+github-actions[bot]@users.noreply.github.com'
+          git add *.svg *.json
+          git commit -m "Autogenerated JaCoCo coverage badges" *.svg *.json
           git push
         fi
 
-#    - name: Create pull request
-#      if: ${{ github.event_name == 'push' }}
-#      uses: peter-evans/create-pull-request@v3.10.0
-#      with:
-#        title: "Autogenerated Jacoco Coverage Badge"
-#        body: > 
-#          Autogenerated Jacoco coverage badge, generated by 
-#          the [jacoco-badge-generator](https://github.com/cicirello/jacoco-badge-generator) 
-#          GitHub action.
-#        commit-message: "Autogenerated Jacoco Coverage Badge"
-#        author: Vincent A Cicirello <cicirello@users.noreply.github.com>
-#        committer: Vincent A Cicirello <cicirello@users.noreply.github.com>
-#        branch: create-pull-request/badge
-#        delete-branch: true
-#        labels: |
-#          automated
+    - name: Comment on PR with coverage percentages
+      if: ${{ github.event_name == 'pull_request' }}
+      run: |
+        REPORT=$(<badges/coverage-summary.json)
+        COVERAGE=$(jq -r '.coverage' <<< "$REPORT")%
+        BRANCHES=$(jq -r '.branches' <<< "$REPORT")%
+        NEWLINE=$'\n'
+        BODY="## JaCoCo Test Coverage Summary Statistics${NEWLINE}* __Coverage:__ ${COVERAGE}${NEWLINE}* __Branches:__ ${BRANCHES}"
+        gh pr comment ${{github.event.pull_request.number}} -b "${BODY}"
+      env:
+        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
