### 2021-11-29 cicirello/InteractiveBinPacking#54 @cicirello
@turketwh javadoc was skipping that class because it has package access, but I'm about to change that with PR #55 since the entire application is in a single Java package there isn't really a reason to isolate any classes with package-level access.
