### 2021-12-05 cicirello/InteractiveBinPacking#56 @turketwh
Feature Req: Statement of Need updates discussing related work, immediate feedback
As I work through reviewing this application as part of a JOSE submission (the review issue thread is linked [here](https://github.com/openjournals/jose-reviews/issues/140) for reference), the *Statement of Need* section in the `paper.md` publication document should be strengthened, particularly with respect to discussing other bin-packing educational software resources and how this software contribution compares.

**Describe the solution you'd like**
Addition of information on other bin packing educational software resources (software which has been published on, or just generally available) should be added, helping to make a case for the contributions of your InteractiveBinPacking software and tutorials. The manuscript on [DraggleSimulator in JOSE](https://jose.theoj.org/papers/10.21105/jose.00070), a software-focused paper similar to yours, provides a example of what this might look like in the first paragraph of its *Statement of Need* section, though I think references to other resources, if they exist, would be helpful (that publication does not cite other work in its *Statement of Need* section).  The manuscript on [BeamBending in Jose](https://jose.theoj.org/papers/10.21105/jose.00065) also is a good example of this type of content in the *Statement of Need* section.

In searches, I have found [this simulator](https://planetcalc.com/917/).

Though not as critical, it would also be useful to reference resources around the value of students getting _immediate_ feedback on their errors, given your software supports this for some of the heuristic approaches. While I don't have a set of resources to share, the SIGCSE proceedings should be a good source of publications focused on automated feedback for computer science-focused content.

Your publication's *Statement of Need* section makes a case for why understanding the bin packing problem and understanding constructive heuristics for the problem are important and for why the interactive aspects of the application are valuable. The above suggestions would help round out this section of the manuscript, helping to show your contribution is novel and employs contemporary pedagogical practices.
