### 2021-12-13 cicirello/InteractiveBinPacking#56 @cicirello
@turketwh Just wanted to acknowledge that I saw this. I'll make the revisions right after end of semester grading is over. Should be straightforward.
