+++ paper/paper.bib
@@ -101,7 +101,7 @@ pages = {120–127}
 @misc{PlanetcalcBinPacking,
 title = {Online Calculator: Bin Packing Problem},
 author = {PlanetCalc},
-howpublished = {https://planetcalc.com/917/},
+url = {https://planetcalc.com/917/},
 year = {2021},
 note = {Accessed: 2021-12-22}
 }
@@ -109,7 +109,7 @@ note = {Accessed: 2021-12-22}
 @misc{Cernohorska2015,
 title = {Heuristics for 1D and 2D bin packing},
 author = {Veronika Černohorská},
-howpublished = {http://www.binpacking.4fan.cz/},
+url = {http://www.binpacking.4fan.cz/},
 year = {2015},
 note = {Czech Technical University in Prague. Accessed: 2021-12-22}
 }
@@ -117,7 +117,7 @@ note = {Czech Technical University in Prague. Accessed: 2021-12-22}
 @misc{GoogleORTools,
 title = {OR-Tools: The Bin Packing Problem},
 author = {{Google Developers}},
-howpublished = {https://developers.google.com/optimization/bin/bin_packing},
+url = {https://developers.google.com/optimization/bin/bin_packing},
 year = {2021},
 month = {August},
 note = {Accessed: 2021-12-22}
