### 2021-12-22 2f8b8ce
added missing urls to citations
