+++ paper/paper.bib
@@ -22,7 +22,8 @@ publisher = {IEEE Computer Society}
   author={Malkevitch, Joseph},
   journal={Feature column from the AMS: Monthly essays on mathematical topics},
   year={2004},
-  month = {June}
+  month = {June},
+  url = {http://www.ams.org/publicoutreach/feature-column/fcarc-packings1}
 }
 
 @article{Malkevitch2004a,
@@ -30,7 +31,8 @@ publisher = {IEEE Computer Society}
   author={Malkevitch, Joseph},
   journal={Feature column from the AMS: Monthly essays on mathematical topics},
   year={2004},
-  month = {May}
+  month = {May},
+  url = {http://www.ams.org/publicoutreach/feature-column/fcarc-bins1}
 }
 
 @article{Wang2006,
