+++ paper/paper.bib
@@ -97,7 +97,7 @@ volume = {23},
 number = {3},
 journal = {Journal of Computing Sciences in Colleges},
 month = {January},
-pages = {120–127}
+pages = {120–-127}
 }
 
 @misc{PlanetcalcBinPacking,
@@ -125,6 +125,55 @@ month = {August},
 note = {Accessed: 2021-12-22}
 }
 
+@inproceedings{Price2017,
+author = {Price, Thomas W. and Dong, Yihuan and Lipovac, Dragan},
+title = {ISnap: Towards Intelligent Tutoring in Novice Programming Environments},
+year = {2017},
+doi = {10.1145/3017680.3017762},
+booktitle = {Proceedings of the 2017 ACM SIGCSE Technical Symposium on Computer Science Education},
+publisher = {ACM},
+pages = {483--488}
+}
+
+@inproceedings{Kim2017,
+author = {Kim, Ada S. and Ko, Amy J.},
+title = {A Pedagogical Analysis of Online Coding Tutorials},
+year = {2017},
+publisher = {ACM},
+doi = {10.1145/3017680.3017728},
+booktitle = {Proceedings of the 2017 ACM SIGCSE Technical Symposium on Computer Science Education},
+pages = {321--326}
+}
+
+@inproceedings{Krusche2018,
+author = {Krusche, Stephan and Seitz, Andreas},
+title = {ArTEMiS: An Automatic Assessment Management System for Interactive Learning},
+year = {2018},
+publisher = {ACM},
+doi = {10.1145/3159450.3159602},
+booktitle = {Proceedings of the 49th ACM Technical Symposium on Computer Science Education},
+pages = {284--289}
+}
+
+@inproceedings{Dicheva2019,
+author = {Dicheva, Darina and Irwin, Keith and Dichev, Christo},
+title = {OneUp: Engaging Students in a Gamified Data Structures Course},
+year = {2019},
+publisher = {ACM},
+doi = {10.1145/3287324.3287480},
+booktitle = {Proceedings of the 50th ACM Technical Symposium on Computer Science Education},
+pages = {386--392}
+}
+
+@inproceedings{Deb2017,
+author = {Deb, Debzani and Fuad, Mohammad Muztaba and Kanan, Mallek},
+title = {Creating Engaging Exercises With Mobile Response System (MRS)},
+year = {2017},
+publisher = {ACM},
+doi = {10.1145/3017680.3017793},
+booktitle = {Proceedings of the 2017 ACM SIGCSE Technical Symposium on Computer Science Education},
+pages = {147--152}
+}
 
 
 
