### 2021-12-22 cicirello/InteractiveBinPacking#56 @cicirello
@turketwh PR #57 includes revisions to the Statement of Need section. I'll wait until you've had a chance to review the changes before I close the issue.
