### 2021-12-22 #140 @turketwh
@labarba @cicirello I have (effectively) finished my review of this JOSE submission.  I have left the *Version* box under the *General checks* criteria unchecked, as I believe the version in the repository is 3.1.0, while it was originally submitted as version 3.0 (as indicated at the top of this review issue).

I asked the author to address three concerns during the review process, impacting the documentation around installation and support,  the Java API documentation, and the *Statement of need* in the publication itself.  I am satisfied with the author's addressing of each concern.

I am in favor of *accepting* the submission.

In my opinion, I believe the software, embedded tutorial, example assignments , and accompanying manuscript make a valuable contribution as an open-source educational tool. Bin packing, as an example of a NP-hard optimization problem, and the idea of heuristics appear in multiple computer science and mathematics courses, suggesting broad applicability.  From using the software, I would argue that the number of heuristics covered and the range of bin packing problems that can be initialized allow for repeated, non-trivial student learning experiences and the example assignments provide a template for how to curate such experiences.  From reviewing the code, the software itself appears cleanly written and well tested, and, in my opinion, is in a state that facilitates extension by others if desired.  The software provides capabilities that go beyond other bin packing tutorials that I can find available and it incorporates various pedagogical best practices.
