+++ paper/paper.bib
@@ -97,7 +97,7 @@ volume = {23},
 number = {3},
 journal = {Journal of Computing Sciences in Colleges},
 month = {January},
-pages = {120–-127}
+pages = {120--127}
 }
 
 @misc{PlanetcalcBinPacking,
