### 2021-12-23 7367b5d
added urls to paper bib file
