+++ paper/paper.bib
@@ -75,7 +75,8 @@ volume = {28},
 number = {6},
 journal = {Journal of Computing Sciences in Colleges},
 month = {June},
-pages = {162--168}
+pages = {162--168},
+url = {https://dl.acm.org/doi/abs/10.5555/2460156.2460187}
 }
 
 @article{Cicirello2009,
@@ -86,7 +87,8 @@ volume = {24},
 number = {6},
 journal = {Journal of Computing Sciences in Colleges},
 month = {June},
-pages = {117--123}
+pages = {117--123},
+url = {https://dl.acm.org/doi/abs/10.5555/1529995.1530020}
 }
 
 @article{Cicirello2007,
@@ -97,7 +99,8 @@ volume = {23},
 number = {3},
 journal = {Journal of Computing Sciences in Colleges},
 month = {January},
-pages = {120--127}
+pages = {120--127},
+url = {https://dl.acm.org/doi/abs/10.5555/1295109.1295137}
 }
 
 @misc{PlanetcalcBinPacking,
