+++ .github/workflows/build.yml
@@ -14,10 +14,10 @@ jobs:
 
     steps:
     - name: Checkout
-      uses: actions/checkout@v2
+      uses: actions/checkout@v3
 
     - name: Checkout badges branch to a badges directory nested inside first checkout
-      uses: actions/checkout@v2
+      uses: actions/checkout@v3
       with:
         ref: badges
         path: badges
