+++ .github/workflows/maven-publish.yml
@@ -13,7 +13,7 @@ jobs:
       artifact_name: interactive-bin-packing
 
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v3
 
     - name: Get the release version
       id: get_version
