### 2022-03-30 #140 @labarba
Oops. I forgot the final steps. @cicirello – could you:

- make a tagged release on the repo, and post the version number here?
- make an archive of the submission on Zenodo (or an equivalent service) and post the archive DOI here?
