### 2022-03-30 #140 @cicirello
@labarba Great. I'll create the release and archive with Zenodo tomorrow morning. I'll comment again after I do.
