### 2022-03-31 #140 @cicirello
@labarba the version number is v3.1.1 and the Zenodo DOI is https://doi.org/10.5281/zenodo.6402123
