### 2022-03-31 #140 @cicirello
@labarba the proof looks good.
